+++ app/Http/Controllers/MarcaController.php
@@ -19,20 +19,35 @@ class MarcaController extends Controller
         return response()->json($marca, 201);
     }
 
-    public function show(Marca $marca)
+    public function show($id)
     {
+        $marca = Marca::find($id);
+        if (!$marca) {
+            return response()->json(['message' => 'Marca no encontrada'], 404);
+        }
         return response()->json($marca, 200);
     }
 
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = Marca::find($id);
+        if (!$marca) {
+            return response()->json(['message' => 'Marca no encontrada'], 404);
+        }
+
         $request->validate(['nombre' => 'string|max:255']);
         $marca->update($request->all());
+        
         return response()->json($marca, 200);
     }
 
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = Marca::find($id);
+        if (!$marca) {
+            return response()->json(['message' => 'Marca no encontrada'], 404);
+        }
+
         $marca->delete();
         return response()->json(['message' => 'Marca eliminada con éxito'], 200);
     }
